Heavy rewrite of tile map, pathfinding, maps, collision, entities and their positioning

+++ Shared/Templates.hpp
@@ -27,8 +27,8 @@ struct ObjectTemplate
     uint32 Entry;
     std::string Name;
     std::string Tileset;
-    uint16 TextureX;
-    uint16 TextureY;
+    uint8 TextureX;
+    uint8 TextureY;
 };
 
 struct WorldObjectTemplate : ObjectTemplate
